Created DPLL.js
Boolean satisfiability factor with a touch of caching mechanism and processing efficiency gains. Hope this can grow into something more feasible to reduce cpu usage but the flaw is not the code presented. CPU is flawed logic but may try to explain those flaws later. In terms of 0, 1 and 0/1 where 0/1 needs 1/0 and 0/1 means same as 0 compared to 1 should be if 0 and 1 is {} and the results of 0 would be then store all 0 before 1 and 1 should result the same in all 1 before 0 while the 0/1 shouldn't exist as its only storing the value of 0 or 1 to have a better cache result. ?!? whatever I can't explain it as I feels flawed.